Moved code from __init__.py
